Voor umit en Svetlana: Nodige klassen zijn Serialised en kunnen worden gelezen! Veel plezier met het xml parsen :D
