+++ src/containing/Container.java
@@ -9,13 +9,16 @@ import com.jme3.bounding.BoundingVolume;
 import com.jme3.material.Material;
 import com.jme3.scene.Node;
 import com.jme3.scene.Spatial;
+import java.io.Serializable;
 
 /**
  *
  * @author account
  */
-public class Container extends Node{
+public class Container extends Node implements Serializable{
 
+    private static final long serialVersionUID = 6529685098267757690L;
+    
     private String id;
     private AssetManager assetManager;
     private Spatial container;
